:recycle: move deprecation utilities to private utils and update references
