+++ discord/appinfo.py
@@ -27,6 +27,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from .utils.private import warn_deprecated
 from .utils.private import get_as_snowflake
 from . import utils
 from .asset import Asset
@@ -262,7 +263,7 @@ class AppInfo:
         .. versionadded:: 1.3
         .. deprecated:: 2.7
         """
-        utils.warn_deprecated(
+        warn_deprecated(
             "summary",
             "description",
             reference="https://discord.com/developers/docs/resources/application#application-object-application-structure",
