+++ discord/commands/core.py
@@ -68,7 +68,8 @@ from ..object import Object
 from ..role import Role
 from ..threads import Thread
 from ..user import User
-from ..utils import MISSING, async_all, find, maybe_coroutine, utcnow, warn_deprecated
+from ..utils import MISSING, async_all, find, maybe_coroutine, utcnow
+from ..utils.private import warn_deprecated
 from .context import ApplicationContext, AutocompleteContext
 from .options import Option, OptionChoice
